Change empty value check to accept zero strings
By using empty() to guard against emptiness, DataHelper::checkForEmptyValue
would previously reject a numeric zero, because trim() would turn it into a
string of zero, which is treated like an empty value.

Now, we harness the fact that trim() already does string coercion, so we can
just to a strict comparison with "".